# Implemented shared_ptr for event pipeline