--- a/src/engine/source/builder/outputs/file.hpp
+++ b/src/engine/source/builder/outputs/file.hpp
@@ -66,7 +66,7 @@ public:
      */
     void write(const types::Event & e)
     {
-        this->m_os << e.str() << std::endl;
+        this->m_os << e->str() << std::endl;
     }
 };