--- a/src/engine/source/server/protocolHandler.hpp
+++ b/src/engine/source/server/protocolHandler.hpp
@@ -47,7 +47,7 @@ public:
      *
      * @return json::Document
      */
-    static json::Document parse(const std::string & event);
+    static std::shared_ptr<json::Document> parse(const std::string & event);
 
     /**
      * @brief process the chunk of data and send messages to dst when. Return