--- a/src/engine/test/source/builder/builders/assetBuilderRuleTest.cpp
+++ b/src/engine/test/source/builder/builders/assetBuilderRuleTest.cpp
@@ -130,6 +130,6 @@ TEST(AssetBuilderRule, BuildsOperates)
     ASSERT_EQ(expected.size(), 2);
     for (auto e : expected)
     {
-        ASSERT_STREQ(e.get("/mapped/field")->GetString(), "value");
+        ASSERT_STREQ(e->get("/mapped/field")->GetString(), "value");
     }
 }