|Related issue|
|---|
|https://github.com/wazuh/wazuh/issues/12612|

In this PR, we change of the observable's JSON event by a shared pointer, so that it is not copied during its life cycle.