@juliancnn 2022-03-11
## PR Reports

### T1 - Benchmark report 
Configuration:
Rate:         	    infinite events/sec
Time:         	          60 sec
Concurrent connections: 1
Dataset: ./test_logs.txt
Output:  /tmp/filepath.txt

Results:
Duration:          60.046406 seconds
Sent events:         1706565
Processed events:    1706565
Lost events:               0


### T2 - Benchmark report
Configuration:
Rate:         	    infinite events/sec
Time:         	          60 sec
Concurrent connections: 1
Dataset: ./test_logs.txt
Output:  /tmp/filepath.txt

Results:
Duration:          60.015869 seconds
Sent events:         1713040
Processed events:    1713040
Lost events:               0


### T4 - Benchmark report

Configuration:
Rate:         	    infinite events/sec
Time:         	          60 sec
Concurrent connections: 1
Dataset: ./test_logs.txt
Output:  /tmp/filepath.txt

Results:
Duration:          60.283485 seconds
Sent events:         1735370
Processed events:    1735370
Lost events:               0


### T8 - Benchmark report
Configuration:
Rate:         	    infinite events/sec
Time:         	          60 sec
Concurrent connections: 1
Dataset: ./test_logs.txt
Output:  /tmp/filepath.txt

Results:
Duration:          60.057701 seconds
Sent events:         1734250
Processed events:    1734250
Lost events:               0


Results Master VS PR:

![image](https://user-images.githubusercontent.com/1891958/157896220-46e3610c-642a-435a-b44d-9434f4fe2d04.png)

![image](https://user-images.githubusercontent.com/1891958/157896272-eea738fb-b4f2-423c-8461-d13b7bf55db8.png)

![image](https://user-images.githubusercontent.com/1891958/157896392-3b6f553c-f7a1-49dd-a2e5-7cffd998e03f.png)

![image](https://user-images.githubusercontent.com/1891958/157896487-32c52761-c476-4379-8a01-ffcfa38c7653.png)


[Reports](https://github.com/wazuh/wazuh/files/8233167/Engine.PR.vs.Master.ods)